+++ includes/class-nuvei-pfw-refund.php
@@ -20,12 +20,16 @@ class Nuvei_Pfw_Refund extends Nuvei_Pfw_Request {
 			Nuvei_Pfw_Logger::write( $data, 'Nuvei_Pfw_Refund error missing mandatoriy parameters.' );
 			return false;
 		}
+        
+        // check if we already have the Order
+        if (empty($this->sc_order)) {
+            $this->sc_order = wc_get_order( $data['order_id'] );
+        }
 
-		$time       = gmdate( 'YmdHis', time() );
-		$order      = wc_get_order( $data['order_id'] );
-		$notify_url = Nuvei_Pfw_String::get_notify_url( $this->plugin_settings );
-		$nuvei_data = $order->get_meta( NUVEI_PFW_TRANSACTIONS );
-		$last_tr    = $this->get_last_transaction( $nuvei_data, array( 'Sale', 'Settle' ) );
+		$time           = gmdate( 'YmdHis', time() );
+		$notify_url     = Nuvei_Pfw_String::get_notify_url( $this->plugin_settings );
+		$nuvei_data     = $this->sc_order->get_meta( NUVEI_PFW_TRANSACTIONS );
+		$last_tr        = $this->get_last_transaction( $nuvei_data, array( 'Sale', 'Settle' ) );
 
 		if ( empty( $last_tr['transactionId'] ) ) {
 			wp_send_json(
@@ -64,13 +68,11 @@ class Nuvei_Pfw_Refund extends Nuvei_Pfw_Request {
 		if ( $order_id < 1 ) {
 			Nuvei_Pfw_Logger::write( $order_id, 'create_refund_request() Error - Post parameter is less than 1.' );
 
-			wp_send_json(
-				array(
-					'status'    => 0,
-					'msg'       => __( 'Post parameter is less than 1.', 'nuvei-payments-for-woocommerce' ),
-					'data'      => array( $order_id ),
-				)
-			);
+			wp_send_json(array(
+                'status'    => 0,
+                'msg'       => __( 'Post parameter is less than 1.', 'nuvei-payments-for-woocommerce' ),
+                'data'      => array( $order_id ),
+            ));
 			exit;
 		}
 
@@ -78,15 +80,10 @@ class Nuvei_Pfw_Refund extends Nuvei_Pfw_Request {
 
 		// error
 		if ( $ref_amount < 0 ) {
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => __(
-						'Invalid Refund amount.',
-						'nuvei-payments-for-woocommerce'
-					),
-				)
-			);
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => __('Invalid Refund amount.', 'nuvei-payments-for-woocommerce'),
+            ));
 			exit;
 		}
 
@@ -94,37 +91,40 @@ class Nuvei_Pfw_Refund extends Nuvei_Pfw_Request {
 
 		// error
 		if ( ! $this->sc_order ) {
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => __( 'Error when try to get the Order.', 'nuvei-payments-for-woocommerce' ),
-				)
-			);
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => __( 'Error when try to get the Order.', 'nuvei-payments-for-woocommerce' ),
+            ));
 			exit;
 		}
-
-		$msg    = '';
-		$resp   = $this->process(
-			array(
-				'order_id'     => $order_id,
-				'ref_amount'   => $ref_amount,
-			)
-		);
+        
+        $current_ord_status = $this->sc_order->get_status();
+        $pending_status     = $this->nuvei_gw->get_option( 'status_pending' );
+        $msg                = '';
+        
+        // first set Pending status
+        $this->sc_order->update_status( $this->nuvei_gw->get_option( 'status_pending' ) );
+        
+        // then create the Refund request
+		$resp = $this->process(array(
+            'order_id'     => $order_id,
+            'ref_amount'   => $ref_amount,
+        ));
 
 		// error
 		if ( false === $resp ) {
 			$msg = __( 'The REST API retun false.', 'nuvei-payments-for-woocommerce' );
 
+            // revert the old status
+            $this->sc_order->update_status( $current_ord_status );
 			$this->sc_order->add_order_note( $msg );
 			$this->sc_order->save();
 
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => $msg,
-				)
-			);
-			exit;
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => $msg,
+            ));
+			wp_die();
 		}
 
 		$json_arr = $resp;
@@ -137,121 +137,124 @@ class Nuvei_Pfw_Refund extends Nuvei_Pfw_Request {
 		if ( ! is_array( $json_arr ) ) {
 			$msg = __( 'Invalid API response.', 'nuvei-payments-for-woocommerce' );
 
+            // revert the old status
+            $this->sc_order->update_status( $current_ord_status );
 			$this->sc_order->add_order_note( $msg );
 			$this->sc_order->save();
 
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => $msg,
-				)
-			);
-			exit;
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => $msg,
+            ));
+			wp_die();
 		}
 
 		// APPROVED
 		if ( ! empty( $json_arr['transactionStatus'] ) && 'APPROVED' == $json_arr['transactionStatus'] ) {
-			// change order status
-			$this->sc_order->update_status( $this->nuvei_gw->get_option( 'status_pending' ) );
+//			// change order status
+//			$this->sc_order->update_status( $this->nuvei_gw->get_option( 'status_pending' ) );
 
 			// save the Refund into transactions, but without status, unitl DMN come
-			unset( $json_arr['status'] );
-			$json_arr['transactionType'] = 'Credit';
-
-			$this->save_transaction_data( $json_arr );
+//			unset( $json_arr['status'] );
+//			$json_arr['transactionType'] = 'Credit';
+//
+//			$this->save_transaction_data( $json_arr );
 
 			wp_send_json( array( 'status' => 1 ) );
-			exit;
+			wp_die();
 		}
 
 		// in case we have message but without status
 		if ( ! isset( $json_arr['status'] ) && isset( $json_arr['msg'] ) ) {
 			$msg = __( 'Refund request problem: ', 'nuvei-payments-for-woocommerce' ) . $json_arr['msg'];
 
+            // revert the old status
+            $this->sc_order->update_status( $current_ord_status );
 			$this->sc_order->add_order_note( $msg );
 			$this->sc_order->save();
 
 			Nuvei_Pfw_Logger::write( $msg );
 
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => $msg,
-				)
-			);
-			exit;
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => $msg,
+            ));
+			wp_die();
 		}
 
 		// the status of the request is ERROR
 		if ( isset( $json_arr['status'] ) && 'ERROR' === $json_arr['status'] ) {
 			$msg = __( 'Request ERROR: ', 'nuvei-payments-for-woocommerce' ) . $json_arr['reason'];
 
+            // revert the old status
+            $this->sc_order->update_status( $current_ord_status );
+            
 			Nuvei_Pfw_Logger::write( $msg );
 
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => $msg,
-				)
-			);
-			exit;
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => $msg,
+            ));
+			wp_die();
 		}
 
 		// the status of the request is SUCCESS, check the transaction status
 		if ( isset( $json_arr['transactionStatus'] ) && 'ERROR' === $json_arr['transactionStatus'] ) {
 			if ( isset( $json_arr['gwErrorReason'] ) && ! empty( $json_arr['gwErrorReason'] ) ) {
 				$msg = $json_arr['gwErrorReason'];
-			} elseif ( isset( $json_arr['paymentMethodErrorReason'] ) && ! empty( $json_arr['paymentMethodErrorReason'] ) ) {
+			} elseif ( isset( $json_arr['paymentMethodErrorReason'] ) 
+                && ! empty( $json_arr['paymentMethodErrorReason'] )
+            ) {
 				$msg = $json_arr['paymentMethodErrorReason'];
 			} else {
 				$msg = __( 'Transaction error.', 'nuvei-payments-for-woocommerce' );
 			}
 
+            // revert the old status
+            $this->sc_order->update_status( $current_ord_status );
 			$this->sc_order->add_order_note( $msg );
 			$this->sc_order->save();
 
 			Nuvei_Pfw_Logger::write( $msg );
 
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => $msg,
-				)
-			);
-			exit;
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => $msg,
+            ));
+			wp_die();
 		}
 
 		if ( isset( $json_arr['transactionStatus'] ) && 'DECLINED' === $json_arr['transactionStatus'] ) {
 			$msg = __( 'The refund was declined.', 'nuvei-payments-for-woocommerce' );
 
+            // revert the old status
+            $this->sc_order->update_status( $current_ord_status );
 			$this->sc_order->add_order_note( $msg );
 			$this->sc_order->save();
 
 			Nuvei_Pfw_Logger::write( $msg );
 
-			wp_send_json(
-				array(
-					'status' => 0,
-					'msg' => $msg,
-				)
-			);
-			exit;
+			wp_send_json(array(
+                'status' => 0,
+                'msg' => $msg,
+            ));
+			wp_die();
 		}
 
 		$msg = __( 'The status of Refund request is UNKONOWN.', 'nuvei-payments-for-woocommerce' );
 
+        // revert the old status
+        $this->sc_order->update_status( $current_ord_status );
 		$this->sc_order->add_order_note( $msg );
 		$this->sc_order->save();
 
 		Nuvei_Pfw_Logger::write( $msg );
 
-		wp_send_json(
-			array(
-				'status' => 0,
-				'msg' => $msg,
-			)
-		);
-		exit;
+		wp_send_json(array(
+            'status' => 0,
+            'msg' => $msg,
+        ));
+		wp_die();
 	}
 
 	protected function get_checksum_params() {