+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Use proper way to get Order private meta data in create_wc_subscr_order();
     * Replaced local Simply Connect with a link to its SDN.
     * Typecast all parameters passed to rtrim() and trim() methods.
+    * When the merchant trigger Refund, change Order status before Nuvei request. In any negative case revert the previous status of the Order.
 ```
 
 # 3.2.1